External OutputManager
